chore(wren-ui): extract error messages collapse to common component
